Add support for WASI targets to Lua sources
This commit is an attempt to work towards mlua-rs/mlua#366 and support
WASI targets in the `mlua` crate. This requires that the Lua source code
can be compiled to WASI targets such as `wasm32-wasip{1,2}`. The C
toolchain used for this is [wasi-sdk] and does not support all that Lua
requires out-of-the-box. This necessitates some edits to Lua sources to
exclude exposing unsupported functions to scripts.

This additionally updates CI to test the `wasm32-wasip2` target inside
of Wasmtime. This currently requires a `dev` release of Wasmtime (builds
from `main`) but the Wasmtime 37.0.0 release next week will suffice once
it's available due to WebAssembly exception-handling support. I've also
dropped the Emscripten tests here as CI wasn't actually testing
emscripten (it forgot `--target`) and the tests are otherwise broken if
re-enabled. Installation of dependencies on CI has additionally been
refactored to a shared "composite" action between the build/test
workflows to avoid duplication between them.

[wasi-sdk]: https://github.com/WebAssembly/wasi-sdk